+++ packages/rsc/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/rsc
 
+## 3.0.0-beta.75
+
+### Patch Changes
+
+- Updated dependencies [e311194]
+  - ai@7.0.0-beta.74
+  - @ai-sdk/provider@4.0.0-beta.9
+  - @ai-sdk/provider-utils@5.0.0-beta.15
+
 ## 3.0.0-beta.74
 
 ### Patch Changes